Add correct black index.js file this time from Turing
