--- prototypes/datasets/index.js
@@ -1,972 +1,0 @@
-
-const { kitties } = require('./datasets/kitties');
-const { clubs } = require('./datasets/clubs');
-const { mods } = require('./datasets/mods');
-const { cakes } = require('./datasets/cakes');
-const { classrooms } = require('./datasets/classrooms');
-const { breweries } = require('./datasets/breweries');
-const { nationalParks } = require('./datasets/nationalParks');
-const { instructors, cohorts } = require('./datasets/turing');
-const { bosses, sidekicks } = require('./datasets/bosses');
-const { constellations, stars } = require('./datasets/astronomy');
-const { weapons, characters } = require('./datasets/ultima');
-const { dinosaurs, humans, movies } = require('./datasets/dinosaurs');
-
-
-
-
-
-
-// SINGLE DATASETS
-// =================================================================
-
-// DATASET: kitties from ./datasets/kitties
-const kittyPrompts = {
-
-orangeKittyNames() {
-  let orangeKits = kitties.filter(kitty => {
-    return kitty.color === 'orange';
-  });
-
-  let orangeNames = [];
-  orangeKits.forEach(kitty => {
-    return orangeNames.push(kitty.name);
-  });
-  // let orangeNames = orangeKits.name;
-  // Return an array of just the names of kitties who are orange e.g.
-  // ['Tiger', 'Snickers']
-  const result = orangeNames;
-  return result;
-
-  // Annotation:
-  // In the first function orangeKits we filter over the array of Kitties,
-  // holding 4 kitty objects. We find the kitties that match our conditional
-  // color === 'orange' and return a new array of the two objects that
-  // match. Next we filter over the new array and forEach element in there
-  // we push the kitty.name into an array and return only the name.
-},
-
-sortByAge() {
-  // Sort the kitties by their age
-  let orderKitAge = kitties.sort((a,b) => a.age < b.age ? 1 : -1);
-
-  const result = orderKitAge;
-  return result;
-
-  // Annotation:
-  // When we use sort() we give it two parameters a and b
-  // we compare a.age to b.age (or the kitty age once we pass in
-  // as the argument) and if age a < b we return 1, which gives
-  // precedence in the array to that object, else we return -1
-  // which does the opposite
-},
-
-growUp() {
-  // Return an array of kitties who have all grown up by 2 years e.g.
-  // [{
-  //   name: 'Felicia',
-  //   age: 4,
-  //   color: 'grey'
-  // },
-  // {
-  //   name: 'Tiger',
-  //   age: 7,
-  //   color: 'orange'
-  // },
-  // ...etc]
-
-  let grownKitts = kitties.map(kitty => {
-    return kitty.age += 2;
-  });
-
-  const result = kitties;
-  return result;
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: clubs from ./datasets/clubs
-const clubPrompts = {
-membersBelongingToClubs() {
-  // Create an object whose keys are the names of people, and whose values are
-  // arrays that include the names of the clubs that person is a part of. e.g.
-  // {
-  //   Louisa: ['Drama', 'Art'],
-  //   Pam: ['Drama', 'Art', 'Chess'],
-  //   ...etc
-  // }
-  //create keys with the name of the member
-  //sort their clubs into an array with their name
-
-  let sortMember = clubs.reduce((acc, club) => {
-    club.members.forEach(member => {
-      if (!acc[member]) {
-        acc[member] = [];
-      }
-      acc[member].push(club.club);
-    });
-    return acc;
-  }, {});
-
-  const result = sortMember;
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: mods from ./datasets/mods
-const modPrompts = {
-studentsPerMod() {
-  // Return an array of objects where the keys are mod (the number of the module)
-  // and studentsPerInstructor (how many students per instructor there are for that mod) e.g.
-  // [
-  //   { mod: 1, studentsPerInstructor: 9 },
-  //   { mod: 2, studentsPerInstructor: 11 },
-  //   { mod: 3, studentsPerInstructor: 10 },
-  //   { mod: 4, studentsPerInstructor: 8 }
-  // ]
-  let total, totalPerInstructor;
-  let classList = [];
-  let studentMath = mods.forEach(mod => {
-    total = (mod.students / mod.instructors);
-    totalPerInstructor = {
-      'mod': mod.mod,
-      'studentsPerInstructor': total
-    };
-    classList.push(totalPerInstructor);
-  });
-
-  const result = classList;
-  return result;
-
-  // Annotation:
-  //For the solution we have to start by looping
-  //over the array of objects and creating a new
-  //variable with the value of the students
-  //divided by the instructors for each object.
-  //Then we create a new object with the new key studentsPerInstructor
-  //and assign it to the new variable.
-  //Finally it wants an array so we push each of those new
-  //objects into an array and return the whole array.
-  //This all happens within thw forEach because the value
-  //for total changes with each object
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: cakes from ./datasets/cakes
-const cakePrompts = {
-stockPerCake() {
-  // Return an array of objects that include just the flavor of the cake and how
-  // much of that cake is in stock e.g.
-  // [
-  //    { flavor: 'dark chocolate', inStock: 15 },
-  //    { flavor: 'yellow', inStock: 14 },
-  //    ..etc
-  // ]
-
-  // let cakeList = [];
-  // let stockPerCake = cakes.forEach(cake => {
-  //   let cakeVariety = {
-  //     "flavor": cake.cakeFlavor,
-  //     "inStock": cake.inStock
-  //   }
-  //   cakeList.push(cakeVariety);
-  // });
-
-  let stockCake = cakes.reduce((accumulator, cake) => {
-    let cakeVariety = {
-      'flavor': cake.cakeFlavor,
-      'inStock': cake.inStock
-    };
-    accumulator.push(cakeVariety);
-    return accumulator;
-  }, []);
-
-  const result = stockCake;
-  return result;
-
-  // Annotation:
-  // A forEach method loops over every object in the cake array.
-  //It creates a new object for each cake with the values of cakeFlavor
-  //and cakeInstock count.
-  //Next we push each new object into an array and return the array.
-  //update: reformatted to reduce with an array accumulator.
-},
-
-onlyInStock() {
-  // Return an array of only the cakes that are in stock
-  // e.g.
-  // [
-  //   {
-  //   cakeFlavor: 'dark chocolate',
-  //   filling: null,
-  //   frosting: 'dark chocolate ganache',
-  //   toppings: ['dutch process cocoa', 'toasted sugar', 'smoked sea salt'],
-  //   inStock: 15
-  // },
-  // {
-  //   cakeFlavor: 'yellow',
-  //   filling: 'citrus glaze',
-  //   frosting: 'chantilly cream',
-  //   toppings: ['berries', 'edible flowers'],
-  //   inStock: 14
-  // },
-  // ..etc
-  // ]
-
-  let availableCakes = cakes.filter(cake => {
-    return cake.inStock >= 1;
-  });
-
-  const result = availableCakes;
-  return result;
-
-  // Annotation:
-  //Using filter, I find all the cakes with an inStock
-  //value greater than or equal to 1. I return only the
-  //cake objects that meet that condition.
-
-},
-
-totalInventory() {
-  // Return the total amount of cakes in stock e.g.
-  // 59
-  // let total = 0;
-  let availableCakes = cakes.reduce((acc, cake) => {
-    acc = cake.inStock + acc;
-    return acc;
-  }, 0);
-
-
-
-  const result = availableCakes;
-  return result;
-
-  // Annotation:
-  //I use filter to iterate over each cake object
-  //in the cakes array, and take the inStock value
-  //and assign it to total, and then add it again and
-  //reassign total, until it goes through all the elements.
-  //Then return total amount
-},
-
-allToppings() {
-  // Return an array of all unique toppings (no duplicates) needed to bake
-  // every cake in the dataset e.g.
-  // ['dutch process cocoa', 'toasted sugar', 'smoked sea salt', 'berries', ..etc]
-
-  let allToppings = cakes.reduce((acc, cake) => {
-    cake.toppings.forEach(topping => {
-      if(!acc.includes(topping)) {
-        acc.push(topping);
-      }
-    });
-    return acc;
-  }, []);
-
-
-  const result = allToppings;
-  return result;
-
-  // Annotation:
-  //First check through the toppings array on each cake
-  //and using the conditional see if the acc already contains that
-  //topping, and if not push it into the array created with reduce
-
-},
-
-groceryList() {
-  // I need to make a grocery list. Please give me an object where the keys are
-  // each topping, and the values are the amount of that topping I need to buy e.g.
-  // {
-  //    'dutch process cocoa': 1,
-  //    'toasted sugar': 3,
-  //    'smoked sea salt': 3,
-  //    'berries': 2,
-  //    ...etc
-  // }
-
-  let groceryList = cakes.reduce((acc, cake) => {
-    cake.toppings.forEach(topping => {
-      if(!acc[topping]) {
-        acc[topping] = 0;
-      }
-      acc[topping]++;
-    });
-    return acc;
-  }, {});
-
-
-  const result = groceryList;
-  return result;
-
-  // Annotation:
-  //When going through the cakes array of objects and
-  //looking at the key toppings. Check first if the acc
-  //object already contains a key with the name of that
-  //topping, if not create it and assign it to 0.
-  //Then always increase the number and reassin it by 1
-  //for each time the topping shows up.
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: classrooms from ./datasets/classrooms
-const classPrompts = {
-feClassrooms() {
-  // Create an array of just the front-end classrooms. e.g.
-  // [
-  //   { roomLetter: 'A', program: 'FE', capacity: 32 },
-  //   { roomLetter: 'C', program: 'FE', capacity: 27 },
-  //   { roomLetter: 'E', program: 'FE', capacity: 22 },
-  //   { roomLetter: 'G', program: 'FE', capacity: 29 }
-  // ]
-
-  let sortFE = classrooms.filter(classroom => {
-    return classroom.program === 'FE';
-  });
-
-  const result = sortFE;
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-totalCapacities() {
-  // Create an object where the keys are 'feCapacity' and 'beCapacity',
-  // and the values are the total capacity for all classrooms in each program e.g.
-  // {
-  //   feCapacity: 110,
-  //   beCapacity: 96
-  // }
-
-  let totalCapacity = classrooms.reduce((acc, program) => {
-    // classroom.forEach(program => {
-    if(!acc[program]) {
-      acc[program] = 0;
-    }
-    acc = acc[program] + acc;
-    console.log(acc);
-    return acc;
-  }, {});
-
-  const result = totalCapacity;
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-sortByCapacity() {
-  // Return the array of classrooms sorted by their capacity (least capacity to greatest)
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-// DATASET: nationalParks from ./datasets/nationalParks
-
-const nationalParksPrompts = {
-getParkVisitList() {
-  /// Return an object containing the names of which parks I need to visit
-  // and the ones I have already visited eg:
-  // {
-  //   parksToVisit: ["Yellowstone", "Glacier", "Everglades"],
-  //   parksVisited: ["Rocky Mountain", "Acadia", "Zion"]
-  //}
-  let sortParks = nationalParks.reduce((acc, park) => {
-    let visited = [];
-    let toVisit = [];
-    if (park.visited) {
-      parksToVisit.push(park.location);
-    } else if (!park.visited) {
-      visited.push(park.location);
-    }
-    acc = {
-      'parksToVisit': toVisit,
-      'parksVisited': visited
-    };
-    console.log(toVisit);
-    return acc;
-  }, {});
-
-  const result = sortParks;
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-getParkInEachState() {
-  // Return an array of objects where the key is the state and the value is its National Park
-  // eg: [ { Colorado: 'Rocky Mountain' },
-  // { Wyoming: 'Yellowstone' },
-  // { Montana: 'Glacier' },
-  // { Maine: 'Acadia' },
-  // { Utah: 'Zion' },
-  // { Florida: 'Everglades' } ]
-
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-getParkActivities() {
-  // Return an array of all the activities I can do
-  // in a National Park. Make sure to exclude duplicates. eg:
-  // [ 'hiking',
-  //   'shoeshoing',
-  //   'camping',
-  //   'fishing',
-  //   'boating',
-  //   'watching wildlife',
-  //   'cross-country skiing',
-  //   'swimming',
-  //   'bird watching',
-  //   'canyoneering',
-  //   'backpacking',
-  //   'rock climbing' ]
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: breweries from ./datasets/breweries
-const breweryPrompts = {
-getBeerCount() {
-  // Return the total beer count of all beers for every brewery e.g.
-  // 40
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-getBreweryBeerCount() {
-  // Return an array of objects where each object has the name of a brewery
-  // and the count of the beers that brewery has e.g.
-  // [
-  //  { name: 'Little Machine Brew', beerCount: 12 },
-  //  { name: 'Ratio Beerworks', beerCount: 5},
-  // ...etc.
-  // ]
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-findHighestAbvBeer() {
-  // Return the beer which has the highest ABV of all beers
-  // e.g.
-  // { name: 'Barrel Aged Nature\'s Sweater', type: 'Barley Wine', abv: 10.9, ibu: 40 }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DOUBLE DATASETS
-// =================================================================
-
-// DATASET: instructors, cohorts from ./datasets/turing
-const turingPrompts = {
-studentsForEachInstructor() {
-  // Return an array of instructors where each instructor is an object
-  // with a name and the count of students in their module. e.g.
-  // [
-  //  { name: 'Pam', studentCount: 21 },
-  //  { name: 'Robbie', studentCount: 18 }
-  // ]
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-studentsPerInstructor() {
-  // Return an object of how many students per teacher there are in each cohort e.g.
-  // {
-  // cohort1806: 9,
-  // cohort1804: 10.5
-  // }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-modulesPerTeacher() {
-  // Return an object where each key is an instructor name and each value is
-  // an array of the modules they can teach based on their skills. e.g.:
-  // {
-  //     Pam: [2, 4],
-  //     Brittany: [2, 4],
-  //     Nathaniel: [2, 4],
-  //     Robbie: [4],
-  //     Leta: [2, 4],
-  //     Travis: [1, 2, 3, 4],
-  //     Louisa: [1, 2, 3, 4],
-  //     Christie: [1, 2, 3, 4],
-  //     Will: [1, 2, 3, 4]
-  //   }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-curriculumPerTeacher() {
-  // Return an object where each key is a curriculum topic and each value is
-  // an array of instructors who teach that topic e.g.:
-  // {
-  //   html: [ 'Travis', 'Louisa' ],
-  //   css: [ 'Travis', 'Louisa' ],
-  //   javascript: [ 'Travis', 'Louisa', 'Christie', 'Will' ],
-  //   recursion: [ 'Pam', 'Leta' ]
-  // }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: bosses, sidekicks from ./datasets/bosses
-const bossPrompts = {
-bossLoyalty() {
-  // Create an array of objects that each have the name of the boss and the sum
-  // loyalty of all their sidekicks. e.g.:
-  // [
-  //   { bossName: 'Jafar', sidekickLoyalty: 3 },
-  //   { bossName: 'Ursula', sidekickLoyalty: 20 },
-  //   { bossName: 'Scar', sidekickLoyalty: 16 }
-  // ]
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: constellations, stars } from ./datasets/astronomy
-const astronomyPrompts = {
-starsInConstellations() {
-  // Return an array of all the stars that appear in any of the constellations
-  // listed in the constellations object e.g.
-  // [
-  //   { name: 'Rigel',
-  //     visualMagnitude: 0.13,
-  //     constellation: 'Orion',
-  //     lightYearsFromEarth: 860,
-  //     color: 'blue' },
-  //   { name: 'Betelgeuse',
-  //     visualMagnitude: 0.5,
-  //     constellation: 'Orion',
-  //     lightYearsFromEarth: 640,
-  //     color: 'red' }
-  // ]
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-starsByColor() {
-  // Return an object with keys of the different colors of the stars,
-  // whose values are arrays containing the star objects that match e.g.
-  // {
-  //   blue: [{obj}, {obj}, {obj}, {obj}, {obj}],
-  //   white: [{obj}, {obj}],
-  //   yellow: [{obj}, {obj}],
-  //   orange: [{obj}],
-  //   red: [{obj}]
-  // }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-constellationsStarsExistIn() {
-  // Return an array of the names of the constellations that the brightest stars are part of e.g.
-
-  //  [ "Canis Major",
-  //    "Carina",
-  //    "Boötes",
-  //    "Auriga",
-  //    "Orion",
-  //    "Lyra",
-  //    "Canis Minor",
-  //    "The Plow",
-  //    "Orion",
-  //    "The Little Dipper" ]
-
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: charaters, weapons from ./datasets/ultima
-const ultimaPrompts = {
-totalDamage() {
-
-  // Return the sum of the amount of damage for all the weapons that our characters can use
-  // Answer => 113
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-charactersByTotal() {
-
-  // Return the sum damage and total range for each character as an object.
-  // ex: [ { Avatar: { damage: 27, range: 24 }, { Iolo: {...}, ...}
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-};
-
-
-
-
-
-
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-// ---------------------------------------------------------------------------
-
-
-
-
-
-
-// DATASET: dinosaurs, humans, movies from ./datasets/dinosaurs
-const dinosaurPrompts = {
-countAwesomeDinosaurs() {
-  // Return an object where each key is a movie title and each value is the
-  // number of awesome dinosaurs in that movie. e.g.:
-  // {
-  //   'Jurassic Park': 5,
-  //   'The Lost World: Jurassic Park': 8,
-  //   'Jurassic Park III': 9,
-  //   'Jurassic World': 11,
-  //   'Jurassic World: Fallen Kingdom': 18
-  // }
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-averageAgePerMovie() {
-  /* Return an object where each key is a movie director's name and each value is
-      an object whose key is a movie's title and whose value is the average age
-      of the cast on the release year of that movie.
-    e.g.:
-    {
-      'Steven Spielberg':
-        {
-          'Jurassic Park': 34,
-          'The Lost World: Jurassic Park': 37
-        },
-      'Joe Johnston':
-        {
-          'Jurassic Park III': 44
-        },
-      'Colin Trevorrow':
-        {
-          'Jurassic World': 56
-         },
-      'J. A. Bayona':
-        {
-          'Jurassic World: Fallen Kingdom': 59
-        }
-    }
-  */
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-uncastActors() {
-  /*
-  Return an array of objects that contain the names of humans who have not been cast in a Jurassic Park movie (yet), their nationality, and their imdbStarMeterRating. The object in the array should be sorted alphabetically by nationality.
-  e.g.
-    [{
-      name: 'Justin Duncan',
-      nationality: 'Alien',
-      imdbStarMeterRating: 0
-    },
-    {
-      name: 'Karin Ohman',
-      nationality: 'Chinese',
-      imdbStarMeterRating: 0
-    },
-    {
-      name: 'Tom Wilhoit',
-      nationality: 'Kiwi',
-      imdbStarMeterRating: 1
-    }, {
-      name: 'Jeo D',
-      nationality: 'Martian',
-      imdbStarMeterRating: 0
-    }]
-  */
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-},
-
-actorsAgesInMovies() {
-  /*
-  Return an array of objects for each human and the age(s) they were in the movie(s) they were cast in, as an array of age(s). Only include humans who were cast in at least one movie.
-  e.g.
-  [ { name: 'Sam Neill', ages: [ 46, 54 ] },
-    { name: 'Laura Dern', ages: [ 26, 34 ] },
-    { name: 'Jeff Goldblum', ages: [ 41, 45, 63, 66 ] },
-    { name: 'Richard Attenborough', ages: [ 70, 74, 92, 95 ] },
-    { name: 'Ariana Richards', ages: [ 14, 18 ] },
-    { name: 'Joseph Mazello', ages: [ 10, 14 ] },
-    { name: 'BD Wong', ages: [ 33, 55, 58 ] },
-    { name: 'Chris Pratt', ages: [ 36, 39 ] },
-    { name: 'Bryce Dallas Howard', ages: [ 34, 37 ] } ]
-  */
-
-  const result = 'REPLACE WITH YOUR RESULT HERE';
-  return result;
-
-  // Annotation:
-  // Write your annotation here as a comment
-}
-};
-
-module.exports = {
-breweryPrompts,
-turingPrompts,
-clubPrompts,
-bossPrompts,
-classPrompts,
-modPrompts,
-kittyPrompts,
-cakePrompts,
-astronomyPrompts,
-ultimaPrompts,
-nationalParksPrompts,
-dinosaurPrompts
-};
